Rename the models in common
