+++ common/node/common/src/models/IFloorplan.ts
