+++ common/node/common/src/models/ISensor.ts
