+++ common/node/common/src/models/index.ts
@@ -1,6 +1,6 @@
-export * from "./config";
-export * from "./device";
-export * from "./floorplan";
-export * from "./schedule";
-export * from "./sensor";
-export * from "./user";
+export * from "./IConfigFile";
+export * from "./IDevice";
+export * from "./IFloorplan";
+export * from "./ISchedule";
+export * from "./ISensor";
+export * from "./IUser";
